requirements.txt
google-cloud-bigquery
pandas
beautifulsoup4
pandas_gbq
unidecode
openpyxl